move api logic into a new module: `api.rs`
